+++ Tests/SuperwallKitTests/Models/Config/ConfigResponseTests.swift
@@ -424,7 +424,8 @@ let response = #"""
         },
         "entitlements": [
           {
-            "id": "test"
+            "identifier": "test",
+            "type": "SERVICE_LEVEL"
           }
         ]
       },